+++ scripts/end-new-cycle.sh
@@ -14,7 +14,7 @@ fi
 
 cd ..
 
-for template in *-enhanced scientific-thesis-template uni-stuttgart-dissertation-template; do
+for template in *-enhanced scientific-thesis-template uni-stuttgart-dissertation-template markdown-latex-quickstart; do
   echo "$template"
   cd $template